test(logic): split order suggestion core part1 tests under test/orders
Refs #2288

Move move/explore, work targets/build, and valid work tile visibility
cases from the monolithic part1 file into three focused test files
(<=400 lines each) under test/orders/.
